+++ api/src/controllers/young/index.js
@@ -21,8 +21,7 @@ const SessionPhase1 = require("../../models/sessionPhase1");
 const ApplicationModel = require("../../models/application");
 const MissionModel = require("../../models/mission");
 const AuthObject = require("../../auth");
-const MeetingPointModel = require("../../models/meetingPoint");
-const BusModel = require("../../models/bus");
+const LigneDeBusModel = require("../../models/PlanDeTransport/ligneBus");
 const YoungAuth = new AuthObject(YoungObject);
 const {
   uploadFile,
@@ -31,7 +30,6 @@ const {
   inSevenDays,
   isYoung,
   isReferent,
-  updatePlacesBus,
   updatePlacesSessionPhase1,
   translateFileStatusPhase1,
   getCcOfYoung,
@@ -39,6 +37,7 @@ const {
   notifDepartmentChange,
   autoValidationSessionPhase1Young,
   deleteFile,
+  updateSeatsTakenInBusLine,
 } = require("../../utils");
 const { sendTemplate } = require("../../sendinblue");
 const { cookieOptions } = require("../../cookie-options");
@@ -410,47 +409,6 @@ router.put("/:id/validate-mission-phase3", passport.authenticate("young", { sess
   }
 });
 
-router.post("/:id/deplacementPhase1Autonomous", passport.authenticate(["referent", "young"], { session: false, failWithError: true }), async (req, res) => {
-  try {
-    const { error, value } = Joi.object({
-      id: Joi.string().required(),
-    })
-      .unknown()
-      .validate({ ...req.params }, { stripUnknown: true });
-    if (error) {
-      capture(error);
-      return res.status(400).send({ ok: false, code: ERRORS.INVALID_PARAMS });
-    }
-
-    const young = await YoungObject.findById(value.id);
-    if (!young) return res.status(404).send({ ok: false, code: ERRORS.NOT_FOUND });
-    if (isYoung(req.user) && young._id.toString() !== req.user._id.toString()) {
-      return res.status(403).send({ ok: false, code: ERRORS.OPERATION_UNAUTHORIZED });
-    }
-    if (isReferent(req.user) && !canEditYoung(req.user, young)) {
-      return res.status(403).send({ ok: false, code: ERRORS.OPERATION_UNAUTHORIZED });
-    }
-
-    const meetingPoint = await MeetingPointModel.findById(young.meetingPointId);
-    const bus = await BusModel.findById(meetingPoint?.busId);
-
-    young.set({
-      deplacementPhase1Autonomous: "true",
-      meetingPointId: null,
-    });
-    await young.save({ fromUser: req.user });
-
-    // if there is a bus, we need to remove the young from it
-    if (bus) await updatePlacesBus(bus);
-
-    return res.status(200).send({ ok: true, data: young });
-  } catch (error) {
-    capture(error);
-    if (error.code === 11000) return res.status(409).send({ ok: false, code: ERRORS.USER_ALREADY_REGISTERED });
-    return res.status(500).send({ ok: false, code: ERRORS.SERVER_ERROR });
-  }
-});
-
 router.put("/", passport.authenticate("young", { session: false, failWithError: true }), async (req, res) => {
   try {
     const { error, value } = validateYoung(req.body, req.user);
@@ -465,8 +423,8 @@ router.put("/", passport.authenticate("young", { session: false, failWithError: 
     if (!canUpdateYoungStatus({ body: value, current: young })) return res.status(403).send({ ok: false, code: ERRORS.OPERATION_UNAUTHORIZED });
 
     if (value?.department && young?.department && value?.department !== young?.department) {
-      await notifDepartmentChange(value.department, SENDINBLUE_TEMPLATES.young.DEPARTMENT_IN, young);
-      await notifDepartmentChange(young.department, SENDINBLUE_TEMPLATES.young.DEPARTMENT_OUT, young);
+      await notifDepartmentChange(value.department, SENDINBLUE_TEMPLATES.young.DEPARTMENT_IN, young, { previousDepartment: young.department });
+      await notifDepartmentChange(young.department, SENDINBLUE_TEMPLATES.young.DEPARTMENT_OUT, young, { newDepartment: value.department });
     }
 
     if (value?.status === YOUNG_STATUS.WITHDRAWN) {
@@ -509,13 +467,10 @@ router.put("/", passport.authenticate("young", { session: false, failWithError: 
       if (sessionPhase1) await updatePlacesSessionPhase1(sessionPhase1, req.user);
     }
 
-    // if they had a meeting point, we check if we need to update the places taken / left in the bus
-    if (young.meetingPointId) {
-      const meetingPoint = await MeetingPointModel.findById(young.meetingPointId);
-      if (meetingPoint) {
-        const bus = await BusModel.findById(meetingPoint.busId);
-        if (bus) await updatePlacesBus(bus);
-      }
+    // if they had a bus, we check if we need to update the places taken / left in the bus
+    if (young.ligneId) {
+      const bus = await LigneDeBusModel.findById(young.ligneId);
+      if (bus) await updateSeatsTakenInBusLine(bus);
     }
     return res.status(200).send({ ok: true, data: young });
   } catch (error) {
@@ -550,11 +505,12 @@ router.put("/:id/change-cohort", passport.authenticate("young", { session: false
     const { cohort, cohortChangeReason, cohortDetailedChangeReason } = value;
 
     const oldSessionPhase1Id = young.sessionPhase1Id;
-    const oldMeetingPointId = young.meetingPointId;
+    const oldBusId = young.ligneId;
     const oldCohort = young.cohort;
-    if (young.cohort !== cohort && (young.sessionPhase1Id || young.meetingPointId)) {
+    if (young.cohort !== cohort && (young.sessionPhase1Id || young.meetingPointId || young.ligneId)) {
       young.set({ sessionPhase1Id: undefined });
       young.set({ meetingPointId: undefined });
+      young.set({ ligneId: undefined });
     }
 
     // si le volontaire change pour la première fois de cohorte, on stocke sa cohorte d'origine
@@ -584,13 +540,10 @@ router.put("/:id/change-cohort", passport.authenticate("young", { session: false
       if (sessionPhase1) await updatePlacesSessionPhase1(sessionPhase1, req.user);
     }
 
-    // if they had a meetingPoint, we check if we need to update the places taken / left in the bus
-    if (oldMeetingPointId) {
-      const meetingPoint = await MeetingPointModel.findById(oldMeetingPointId);
-      if (meetingPoint) {
-        const bus = await BusModel.findById(meetingPoint.busId);
-        if (bus) await updatePlacesBus(bus);
-      }
+    // if they had a bus, we check if we need to update the places taken / left in the bus
+    if (oldBusId) {
+      const bus = await LigneDeBusModel.findById(oldBusId);
+      if (bus) await updateSeatsTakenInBusLine(bus);
     }
 
     const referents = await ReferentModel.find({ role: ROLES.REFERENT_DEPARTMENT, department: young.department });
@@ -629,46 +582,6 @@ router.put("/:id/change-cohort", passport.authenticate("young", { session: false
         cohortPeriod: translateCohort(cohort),
       },
     });
-
-    res.status(200).send({ ok: true, data: young });
-  } catch (error) {
-    capture(error);
-    res.status(500).send({ ok: false, code: ERRORS.SERVER_ERROR });
-  }
-});
-
-router.post("/:id/session-phase1/cancel", passport.authenticate("referent", { session: false, failWithError: true }), async (req, res) => {
-  try {
-    const { error, value: id } = validateId(req.params.id);
-    if (error) {
-      capture(error);
-      return res.status(400).send({ ok: false, code: ERRORS.INVALID_BODY, error });
-    }
-
-    const young = await YoungObject.findById(id);
-    if (!young) return res.status(404).send({ ok: false, code: ERRORS.YOUNG_NOT_FOUND });
-
-    const oldSessionPhase1Id = young.sessionPhase1Id;
-    const oldMeetingPointId = young.meetingPointId;
-    young.set({ sessionPhase1Id: undefined });
-    young.set({ meetingPointId: undefined });
-
-    await young.save({ fromUser: req.user });
-
-    // if they had a session, we check if we need to update the places taken / left
-    if (oldSessionPhase1Id) {
-      const sessionPhase1 = await SessionPhase1.findById(oldSessionPhase1Id);
-      if (sessionPhase1) await updatePlacesSessionPhase1(sessionPhase1, req.user);
-    }
-
-    // if they had a meetingPoint, we check if we need to update the places taken / left in the bus
-    if (oldMeetingPointId) {
-      const meetingPoint = await MeetingPointModel.findById(oldMeetingPointId);
-      if (meetingPoint) {
-        const bus = await BusModel.findById(meetingPoint.busId);
-        if (bus) await updatePlacesBus(bus);
-      }
-    }
 
     res.status(200).send({ ok: true, data: young });
   } catch (error) {