+++ api/src/templates/convocation/index.js
@@ -5,9 +5,18 @@ const { getSignedUrl, getBaseUrl, sanitizeAll } = require("../../utils");
 const CohesionCenterModel = require("../../models/cohesionCenter");
 const SessionPhase1 = require("../../models/sessionPhase1");
 const MeetingPointModel = require("../../models/meetingPoint");
+const CohortModel = require("../../models/cohort");
 const BusModel = require("../../models/bus");
+
+const LigneBusModel = require("../../models/PlanDeTransport/ligneBus");
+const LigneToPointModel = require("../../models/PlanDeTransport/ligneToPoint");
+const PointDeRassemblementModel = require("../../models/PlanDeTransport/pointDeRassemblement");
+
 const DepartmentServiceModel = require("../../models/departmentService");
 const { formatStringDate, formatStringDateTimezoneUTC } = require("snu-lib");
+
+const datefns = require("date-fns");
+var { fr } = require("date-fns/locale");
 
 const isFromDOMTOM = (young) => {
   return (
@@ -29,48 +38,10 @@ function getBottom() {
   return getSignedUrl("convocation/bottom.png");
 }
 
-// ! WARNING : Change date also in app/src/scenes/phase1/components/Convocation.js
-const departureMeetingDate = {
-  2021: "lundi 20 février, 14:00",
-  "Février 2022": "dimanche 13 février, 16:00",
-  "Juin 2022": "dimanche 12 juin, 16:00",
-  "Juillet 2022": "dimanche 03 juillet, 16:00",
-};
-
-const departureMeetingDateException = {
-  2021: "lundi 20 février, 14:00",
-  "Février 2022": "dimanche 13 février, 16:00",
-  "Juin 2022": "mercredi 15 juin, 10:00",
-  "Juillet 2022": "mercredi 06 juillet, 10:00",
-};
-
-const returnMeetingDate = {
-  2021: "mardi 02 juillet, 14:00",
-  "Février 2022": "vendredi 25 février, 11:00",
-  "Juin 2022": "vendredi 24 juin, 11:00",
-  "Juillet 2022": "vendredi 15 juillet, 11:00",
-};
-
-const COHESION_STAY_DATE_STRING = {
-  2021: "20 février au 02 juillet 2021  ",
-  "Février 2022": "13 février au 25 février 2022",
-  "Juin 2022": "12 juin au 24 juin 2022",
-  "Juillet 2022": "03 juillet au 15 juillet 2022",
-};
-
 const render = async (young) => {
-  const getDepartureMeetingDate = (meetingPoint) => {
-    if (young.deplacementPhase1Autonomous === "true" || !meetingPoint)
-      return young.grade !== "Terminale" ? departureMeetingDate[young.cohort] : departureMeetingDateException[young.cohort]; //new Date("2021-06-20T14:30:00.000+00:00");
-    return meetingPoint.departureAtString;
-  };
-  const getReturnMeetingDate = (meetingPoint) => {
-    if (young.deplacementPhase1Autonomous === "true" || !meetingPoint) return returnMeetingDate[young.cohort]; // new Date("2021-07-02T12:00:00.000+00:00");
-    return meetingPoint.returnAtString;
-  };
   const getMeetingAddress = (meetingPoint, center) => {
     if (young.deplacementPhase1Autonomous === "true" || !meetingPoint) return `${center.address} ${center.zip} ${center.city}`;
-    return meetingPoint.departureAddress;
+    return meetingPoint.address + " " + meetingPoint.zip + " " + meetingPoint.city;
   };
 
   try {
@@ -80,8 +51,18 @@ const render = async (young) => {
     if (!session) throw `session ${young.sessionPhase1Id} not found for young ${young._id}`;
     const center = await CohesionCenterModel.findById(session.cohesionCenterId);
     if (!center) throw `center ${session.cohesionCenterId} not found for young ${young._id} - session ${session._id}`;
-    const meetingPoint = await MeetingPointModel.findById(young.meetingPointId);
-    const bus = await BusModel.findById(meetingPoint?.busId);
+
+    const meetingPoint = await PointDeRassemblementModel.findById(young.meetingPointId);
+    let ligneToPoint = null;
+    let ligneBus = null;
+
+    if (meetingPoint && young.ligneId) {
+      ligneBus = await LigneBusModel.findById(young.ligneId);
+      ligneToPoint = await LigneToPointModel.findOne({ lineId: young.ligneId, meetingPointId: young.meetingPointId });
+    }
+
+    const cohort = await CohortModel.findOne({ name: young.cohort });
+
     const service = await DepartmentServiceModel.findOne({ department: young?.department });
     if (!service) throw `service not found for young ${young._id}, center ${center?._id} in department ${young?.department}`;
     const contacts = service?.contacts.filter((c) => c.cohort === young.cohort) || [];
@@ -105,34 +86,20 @@ const render = async (young) => {
       .replace(/{{ADDRESS}}/g, sanitizeAll(young.address))
       .replace(/{{ZIP}}/g, sanitizeAll(young.zip))
       .replace(/{{CITY}}/g, sanitizeAll(young.city))
-      .replace(/{{COHESION_STAY_DATE_STRING}}/g, sanitizeAll(COHESION_STAY_DATE_STRING[young.cohort]))
+      .replace(
+        /{{COHESION_STAY_DATE_STRING}}/g,
+        sanitizeAll(datefns.format(new Date(cohort?.dateStart), "dd MMMM", { locale: fr }) + " au " + datefns.format(new Date(cohort?.dateEnd), "dd MMMM yyyy", { locale: fr })),
+      )
       .replace(/{{COHESION_CENTER_NAME}}/g, sanitizeAll(center.name))
       .replace(/{{COHESION_CENTER_ADDRESS}}/g, sanitizeAll(center.address))
       .replace(/{{COHESION_CENTER_ZIP}}/g, sanitizeAll(center.zip))
       .replace(/{{COHESION_CENTER_CITY}}/g, sanitizeAll(center.city))
-      .replace(
-        /{{MEETING_DATE}}/g,
-        sanitizeAll(
-          `<b>Le</b> ${getDepartureMeetingDate(meetingPoint)
-            .split(/[,\s]+/)
-            .slice(0, 3)
-            .join(" ")}`,
-        ),
-      )
-      .replace(/{{MEETING_HOURS}}/g, sanitizeAll(`<b>A</b> ${getDepartureMeetingDate(meetingPoint).split(",")[1]}`))
+      .replace(/{{MEETING_DATE}}/g, sanitizeAll(`<b>Le</b> ${datefns.format(new Date(cohort?.dateStart), "EEEE dd MMMM", { locale: fr })}`))
+      .replace(/{{MEETING_HOURS}}/g, sanitizeAll(`<b>A</b> ${meetingPoint ? ligneToPoint.meetingHour : "16:00"}`))
       .replace(/{{MEETING_ADDRESS}}/g, sanitizeAll(`<b>Au</b> ${getMeetingAddress(meetingPoint, center)}`))
-      .replace(/{{TRANSPORT}}/g, sanitizeAll(bus ? `<b>Numéro de transport</b> : ${bus.idExcel}` : ""))
-      .replace(
-        /{{MEETING_DATE_RETURN}}/g,
-        sanitizeAll(
-          getReturnMeetingDate(meetingPoint)
-            .split(/[,\s]+/)
-            .slice(0, 3)
-            .join(" "),
-        ),
-      )
-      .replace(/{{MEETING_HOURS_RETURN}}/g, sanitizeAll(getReturnMeetingDate(meetingPoint).split(",")[1]))
-
+      .replace(/{{TRANSPORT}}/g, sanitizeAll(ligneBus ? `<b>Numéro de transport</b> : ${ligneBus.busId}` : ""))
+      .replace(/{{MEETING_DATE_RETURN}}/g, sanitizeAll(datefns.format(new Date(cohort?.dateEnd), "EEEE dd MMMM", { locale: fr })))
+      .replace(/{{MEETING_HOURS_RETURN}}/g, sanitizeAll(meetingPoint ? ligneToPoint.returnHour : "11:00"))
       .replace(/{{BASE_URL}}/g, sanitizeAll(getBaseUrl()))
       .replace(/{{TOP}}/g, sanitizeAll(getTop()))
       .replace(/{{BOTTOM}}/g, sanitizeAll(getBottom()))
@@ -142,7 +109,7 @@ const render = async (young) => {
   }
 };
 
-// todo ⚠️ not updates because no Février 2022 ⚠️
+// todo ⚠️ not updates because no Février 2023⚠️
 const renderDOMTOM = async (young) => {
   try {
     if (!["Février 2022", "Juin 2022", "Juillet 2022"].includes(young.cohort)) throw `young ${young.id} unauthorized`;
@@ -156,31 +123,33 @@ const renderDOMTOM = async (young) => {
     const contacts = service?.contacts.filter((c) => c.cohort === young.cohort) || [];
 
     const html = fs.readFileSync(path.resolve(__dirname, "./cohesionDOMTOM.html"), "utf8");
-    return html
-      .replace(
-        /{{CONTACTS}}/g,
-        sanitizeAll(
-          contacts
-            .map((contact) => {
-              return `<li>${contact.contactName} - ${contact.contactPhone} - ${contact.contactMail}</li>`;
-            })
-            .join(""),
-        ),
-      )
-      .replace(/{{DATE}}/g, sanitizeAll(formatStringDate(Date.now())))
-      .replace(/{{FIRST_NAME}}/g, sanitizeAll(young.firstName))
-      .replace(/{{LAST_NAME}}/g, sanitizeAll(young.lastName))
-      .replace(/{{BIRTHDATE}}/g, sanitizeAll(formatStringDateTimezoneUTC(young.birthdateAt)))
-      .replace(/{{ADDRESS}}/g, sanitizeAll(young.address))
-      .replace(/{{ZIP}}/g, sanitizeAll(young.zip))
-      .replace(/{{CITY}}/g, sanitizeAll(young.city))
-      .replace(/{{COHESION_STAY_DATE_STRING}}/g, sanitizeAll(COHESION_STAY_DATE_STRING[young.cohort]))
-      .replace(/{{COHESION_CENTER_NAME}}/g, sanitizeAll(center.name))
-      .replace(/{{COHESION_CENTER_ADDRESS}}/g, sanitizeAll(center.address))
-      .replace(/{{COHESION_CENTER_ZIP}}/g, sanitizeAll(center.zip))
-      .replace(/{{COHESION_CENTER_CITY}}/g, sanitizeAll(center.city))
-      .replace(/{{BASE_URL}}/g, sanitizeAll(getBaseUrl()))
-      .replace(/{{GENERAL_BG}}/g, sanitizeAll(getBg()));
+    return (
+      html
+        .replace(
+          /{{CONTACTS}}/g,
+          sanitizeAll(
+            contacts
+              .map((contact) => {
+                return `<li>${contact.contactName} - ${contact.contactPhone} - ${contact.contactMail}</li>`;
+              })
+              .join(""),
+          ),
+        )
+        .replace(/{{DATE}}/g, sanitizeAll(formatStringDate(Date.now())))
+        .replace(/{{FIRST_NAME}}/g, sanitizeAll(young.firstName))
+        .replace(/{{LAST_NAME}}/g, sanitizeAll(young.lastName))
+        .replace(/{{BIRTHDATE}}/g, sanitizeAll(formatStringDateTimezoneUTC(young.birthdateAt)))
+        .replace(/{{ADDRESS}}/g, sanitizeAll(young.address))
+        .replace(/{{ZIP}}/g, sanitizeAll(young.zip))
+        .replace(/{{CITY}}/g, sanitizeAll(young.city))
+        //.replace(/{{COHESION_STAY_DATE_STRING}}/g, sanitizeAll(COHESION_STAY_DATE_STRING[young.cohort]))
+        .replace(/{{COHESION_CENTER_NAME}}/g, sanitizeAll(center.name))
+        .replace(/{{COHESION_CENTER_ADDRESS}}/g, sanitizeAll(center.address))
+        .replace(/{{COHESION_CENTER_ZIP}}/g, sanitizeAll(center.zip))
+        .replace(/{{COHESION_CENTER_CITY}}/g, sanitizeAll(center.city))
+        .replace(/{{BASE_URL}}/g, sanitizeAll(getBaseUrl()))
+        .replace(/{{GENERAL_BG}}/g, sanitizeAll(getBg()))
+    );
   } catch (e) {
     throw e;
   }