+++ app/src/scenes/phase1/components/step/stepMedicalFile.js
@@ -12,6 +12,7 @@ import ModalConfirm from "../../../../components/modals/ModalConfirm";
 import { SENDINBLUE_TEMPLATES } from "../../../../utils";
 import { toastr } from "react-redux-toastr";
 import plausibleEvent from "../../../../services/plausible";
+import { CDN_BASE_URL } from "../../../representants-legaux/commons";
 
 export default function StepMedicalField({ young }) {
   const [stateMobil, setStateMobil] = useState(false);
@@ -39,7 +40,7 @@ export default function StepMedicalField({ young }) {
     const { ok } = await api.post(`/young/${young._id}/email/${SENDINBLUE_TEMPLATES.young.LINK}`, {
       object: `Fiche sanitaire à compléter`,
       message: "Vous trouverez téléchargeable ci-dessous la fiche sanitaire à compléter.",
-      link: "https://cni-bucket-prod.cellar-c2.services.clever-cloud.com/file/fiche_sanitaire_2022.pdf?utm_campaign=transactionnel+telecharger+docum&utm_source=notifauto&utm_medium=mail+410+telecharger",
+      link: CDN_BASE_URL + "/file/fiche-sanitaire-2023.pdf" + "?utm_campaign=transactionnel+telecharger+docum&utm_source=notifauto&utm_medium=mail+410+telecharger",
     });
     if (ok) toastr.success(`Document envoyé à ${young.email}`);
     else toastr.error("Erreur lors de l'envoie du document");
@@ -84,7 +85,7 @@ export default function StepMedicalField({ young }) {
                   <HiOutlineMail className="h-5 w-5 text-gray-600" />
                 </WithTooltip>
               </button>
-              <a target="blank" href="https://cni-bucket-prod.cellar-c2.services.clever-cloud.com/file/fiche_sanitaire_2022.pdf" onClick={handleDownload}>
+              <a target="blank" href={CDN_BASE_URL + "/file/fiche-sanitaire-2023.pdf"} onClick={handleDownload}>
                 <button
                   type="button"
                   className={`flex flex-row items-center justify-center px-4 py-2 rounded-lg ${
@@ -131,7 +132,7 @@ export default function StepMedicalField({ young }) {
                   className="flex flex-row w-full items-center justify-center px-4 py-2 mb-2 rounded-lg bg-blue-600 cursor-pointer hover:scale-105"
                   onClick={handleDownload}>
                   <HiOutlineDownload className="h-5 w-5 text-blue-300 mr-2" />
-                  <a target="blank" href="https://cni-bucket-prod.cellar-c2.services.clever-cloud.com/file/fiche_sanitaire_2022.pdf" onClick={handleDownload}>
+                  <a target="blank" href={CDN_BASE_URL + "/file/fiche-sanitaire-2023.pdf"} onClick={handleDownload}>
                     <span className="text-white text-sm">Télécharger</span>
                   </a>
                 </button>